refactor: move LocalizedField to a separate localization module
