+++ valorant/models/events.py
@@ -26,7 +26,8 @@ from datetime import datetime
 
 from pydantic import Field
 
-from .base import BaseUUIDModel, LocalizedField
+from .base import BaseUUIDModel
+from .localization import LocalizedField
 
 __all__ = ('Event',)
