+++ valorant/models/maps.py
@@ -26,7 +26,8 @@ from typing import Any
 
 from pydantic import Field
 
-from .base import BaseModel, BaseUUIDModel, LocalizedField
+from .base import BaseModel, BaseUUIDModel
+from .localization import LocalizedField
 
 __all__ = (
     'Callout',
