+++ valorant/models/player_titles.py
@@ -24,7 +24,8 @@ DEALINGS IN THE SOFTWARE.
 
 from pydantic import Field
 
-from .base import BaseUUIDModel, LocalizedField
+from .base import BaseUUIDModel
+from .localization import LocalizedField
 
 __all__ = ('PlayerTitle',)
