+++ valorant/models/seasons.py
@@ -31,7 +31,8 @@ from uuid import UUID
 from pydantic import Field
 
 from ..enums import SeasonType
-from .base import BaseUUIDModel, LocalizedField
+from .base import BaseUUIDModel
+from .localization import LocalizedField
 
 __all__ = (
     'Border',
